update to match tutorial output
