+++ circuit/Nargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "noirjs_demo"
+name = "circuit"
 type = "bin"
 authors = [""]
 compiler_version = ">=0.31.0"
