--- package.json
@@ -1,23 +1,0 @@
-{
-  "name": "noirjs_demo",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "vite --open",
-    "build": "vite build",
-    "preview": "vite preview"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@noir-lang/backend_barretenberg": "^0.31.0",
-    "@noir-lang/noir_js": "^0.31.0"
-  },
-  "devDependencies": {
-    "rollup-plugin-copy": "^3.5.0",
-    "vite": "^5.0.0"
-  }
-}
