+++ vite-project/index.html
@@ -15,8 +15,12 @@
   </style>
 </head>
 <body>
-  <script type="module" src="/app.js"></script>
-  <h1>Very basic Noir app</h1>
+  <script type="module" src="/main.js"></script>
+  <h1>Noir app</h1>
+  <div class="input-area">
+    <input id="guessInput" type="number" placeholder="Enter your guess" />
+    <button id="submitGuess">Submit Guess</button>
+  </div>
   <div class="outer">
     <div id="logs" class="inner"><h2>Logs</h2></div>
     <div id="results" class="inner"><h2>Proof</h2></div>
