+++ vite-project/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "vite-project",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "rollup-plugin-copy": "^3.5.0",
+    "vite": "^5.3.4"
+  },
+  "dependencies": {
+    "@noir-lang/backend_barretenberg": "^0.31.0",
+    "@noir-lang/noir_js": "^0.31.0"
+  }
+}
